requirements.txt
fastapi==0.120.0
uvicorn==0.38.0
python-dotenv==1.2.1
sqlalchemy==2.0.44
psycopg2-binary
passlib==1.7.4
bcrypt==3.2.2
pydantic[email]==2.12.3
pyjwt==2.10.1
uuid6==2025.0.1
python-multipart==0.0.20
